[PIP][Connector][Spark] Extract SparkInputPartitionReader from paimon spark connector
