+++ paimon-spark/paimon-spark-common/src/main/java/org/apache/paimon/spark/SparkReaderFactory.java
@@ -51,30 +51,6 @@ public class SparkReaderFactory implements PartitionReaderFactory {
         }
         RecordReaderIterator<InternalRow> iterator = new RecordReaderIterator<>(reader);
         SparkInternalRow row = new SparkInternalRow(readBuilder.readType());
-        return new PartitionReader<org.apache.spark.sql.catalyst.InternalRow>() {
-
-            @Override
-            public boolean next() {
-                if (iterator.hasNext()) {
-                    row.replace(iterator.next());
-                    return true;
-                }
-                return false;
-            }
-
-            @Override
-            public org.apache.spark.sql.catalyst.InternalRow get() {
-                return row;
-            }
-
-            @Override
-            public void close() throws IOException {
-                try {
-                    iterator.close();
-                } catch (Exception e) {
-                    throw new IOException(e);
-                }
-            }
-        };
+        return new SparkInputPartitionReader(iterator, row);
     }
 }
